Extract common data singleton class
